+++ src/pylongfellow/__init__.py
@@ -1,8 +1,8 @@
 """A Python interface to implementations of Longfellow zero-knowledge mdoc proofs.
 
 [`Pylongfellow`][pylongfellow.Pylongfellow] is constructed with a choice of backend
-(`google-cpp` or `isrg-rust`). It provides functions for proving, verifying, and circuit
-generation.
+(`google-cpp` or `isrg-rust`). It proves and verifies.
+[`generate_circuit`][pylongfellow.backends.google_cpp.generate_circuit] generates circuits.
 """
 
 from importlib.metadata import version